LayoutTests/fast/css/content-visibility-invalidate-if-disabled.html

<!DOCTYPE html><!-- webkit-test-runner [ CSSContentVisibilityEnabled=false ] -->
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="target" style="display: none"></div>
<script>
description("Tests that content-visibility is not exposed when the feature is disabled");

shouldBeFalse("'contentVisibility' in document.documentElement.style");
shouldBeFalse("'content-visibility' in getComputedStyle(document.documentElement)");
shouldBeTrue("getComputedStyle(document.getElementById('target')).getPropertyValue('content-visibility') == ''");
shouldBeFalse("CSS.supports('content-visibility: hidden')");
shouldBeFalse("CSS.supports('content-visibility: inherit')");

</script>
</body>
</html>
