LayoutTests/fast/css/content-visibility-invalidate-if-disabled-expected.txt

Tests that content-visibility is not exposed when the feature is disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'contentVisibility' in document.documentElement.style is false
PASS 'content-visibility' in getComputedStyle(document.documentElement) is false
PASS getComputedStyle(document.getElementById('target')).getPropertyValue('content-visibility') == '' is true
PASS CSS.supports('content-visibility: hidden') is false
PASS CSS.supports('content-visibility: inherit') is false
PASS successfullyParsed is true

TEST COMPLETE
